✨ feat(runtime): wire initial_state_seed through AgentLoopParams and BackendDelegateRunRequest
Extract inline test modules for backend/local, backend/mod, loop_runner/orchestrator
to keep parents within the file-length budget.
