+++ crates/awaken-runtime/tests/tool_side_effects.rs
@@ -293,6 +293,7 @@ async fn tool_state_mutation_applied_after_execution() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -341,6 +342,7 @@ async fn tool_scheduled_action_executed() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -427,6 +429,7 @@ async fn tool_empty_command_has_no_side_effects() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -485,6 +488,7 @@ async fn parallel_tool_commands_merge() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
