+++ crates/awaken/examples/live_test.rs
@@ -140,6 +140,7 @@ async fn main() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await;
