+++ crates/awaken/examples/permission_filter_live.rs
@@ -283,6 +283,7 @@ async fn main() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await;
