+++ crates/awaken/examples/tool_call_live.rs
@@ -217,6 +217,7 @@ async fn main() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await;
