+++ crates/awaken/tests/agent_loop.rs
@@ -222,6 +222,7 @@ async fn single_step_natural_end() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -292,6 +293,7 @@ async fn run_level_model_override_selects_upstream_model() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -342,6 +344,7 @@ async fn tool_call_then_response() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -391,6 +394,7 @@ async fn tool_call_state_machine_transitions() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -445,6 +449,7 @@ async fn multiple_tool_calls_in_one_step() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -497,6 +502,7 @@ async fn max_rounds_exceeded() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -555,6 +561,7 @@ async fn unknown_tool_returns_error_result_not_crash() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap(); // Should NOT error — unknown tool produces ToolResult::error
@@ -614,6 +621,7 @@ async fn failing_tool_produces_error_result_continues_loop() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -651,6 +659,7 @@ async fn events_have_correct_sequence_for_single_step() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -727,6 +736,7 @@ async fn events_have_correct_sequence_with_tool_call() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -804,6 +814,7 @@ async fn lifecycle_state_reflects_custom_run_id() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -872,6 +883,7 @@ async fn phase_hooks_fire_during_loop() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -977,6 +989,7 @@ async fn tool_suspension_transitions_run_to_waiting() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1020,6 +1033,7 @@ async fn resume_with_use_decision_as_tool_result() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1070,6 +1084,7 @@ async fn resume_with_use_decision_as_tool_result() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1113,6 +1128,7 @@ async fn resume_with_cancel_marks_tool_cancelled() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1162,6 +1178,7 @@ async fn resume_with_cancel_marks_tool_cancelled() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1203,6 +1220,7 @@ async fn resume_with_replay_tool_call() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1272,6 +1290,7 @@ async fn resume_with_replay_tool_call() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1312,6 +1331,7 @@ async fn resume_with_pass_decision_to_tool() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await;
     // This might not work because tool_call name is "passthrough" but we only have "dangerous".
@@ -1342,6 +1362,7 @@ async fn resume_with_pass_decision_to_tool() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1412,6 +1433,7 @@ async fn resume_with_pass_decision_to_tool() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1442,6 +1464,7 @@ async fn resume_rejects_non_waiting_run() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1494,6 +1517,7 @@ async fn resume_rejects_unknown_call_id() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1627,6 +1651,7 @@ async fn cancel_during_streaming_terminates_run() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1666,6 +1691,7 @@ async fn cancel_before_inference_terminates_immediately() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1721,6 +1747,7 @@ async fn state_snapshot_emitted_after_phase() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -1939,6 +1966,7 @@ async fn frontend_tool_intercept_suspend_and_resume() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2015,6 +2043,7 @@ async fn frontend_tool_intercept_suspend_and_resume() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2059,6 +2088,7 @@ async fn injected_frontend_tool_uses_suspension_id_resume_chain() {
         frontend_tools: vec![frontend_tool.clone()],
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2113,6 +2143,7 @@ async fn injected_frontend_tool_uses_suspension_id_resume_chain() {
         frontend_tools: vec![frontend_tool],
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2207,6 +2238,7 @@ async fn tool_intercept_block_terminates_run() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2338,6 +2370,7 @@ async fn tool_intercept_set_result_skips_execution() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2385,6 +2418,7 @@ async fn suspended_tool_preserves_state_across_resume() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2446,6 +2480,7 @@ async fn suspended_tool_preserves_state_across_resume() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2519,6 +2554,7 @@ async fn decision_channel_resume_resolves_suspended_call() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2580,6 +2616,7 @@ async fn cancel_decision_marks_tool_cancelled() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2625,6 +2662,7 @@ async fn permission_hook_blocks_denied_tool() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2717,6 +2755,7 @@ async fn intercept_suspend_preserves_ticket_resume_mode() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2788,6 +2827,7 @@ async fn intercept_suspend_preserves_ticket_resume_mode() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2843,6 +2883,7 @@ async fn multiple_tool_calls_partial_intercept() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2914,6 +2955,7 @@ async fn intercept_set_result_emits_tool_call_done_event() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -2984,6 +3026,7 @@ async fn prepare_resume_preserves_arguments_and_records_decision_payload() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3045,6 +3088,7 @@ async fn prepare_resume_preserves_arguments_and_records_decision_payload() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3204,6 +3248,7 @@ async fn concurrent_suspend_and_resume_via_channel() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3357,6 +3402,7 @@ async fn single_tool_call_can_suspend_multiple_times_via_decision_channel() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     });
 
     let (result, ()) = tokio::join!(run, sender);
@@ -3436,6 +3482,7 @@ async fn tool_call_lifecycle_complete_transitions_in_loop() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3507,6 +3554,7 @@ async fn tool_call_lifecycle_complete_transitions_in_loop() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3568,6 +3616,7 @@ async fn parallel_tools_one_fails_other_succeeds() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3632,6 +3681,7 @@ async fn sequential_tools_stop_after_first_suspension() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3698,6 +3748,7 @@ async fn stop_policy_max_rounds_terminates() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3771,6 +3822,7 @@ async fn cancel_during_tool_execution() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3816,6 +3868,7 @@ async fn empty_tool_calls_natural_end() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -3927,6 +3980,7 @@ async fn context_message_injected_before_inference() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4015,6 +4069,7 @@ async fn tool_execution_preserves_arguments() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4169,6 +4224,7 @@ async fn retry_startup_error_propagates() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await;
 
@@ -4214,6 +4270,7 @@ async fn inference_request_uses_configured_upstream_model() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4273,6 +4330,7 @@ async fn truncation_with_tool_calls_no_retry() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4379,6 +4437,7 @@ async fn truncation_recovery_exhausts_retries() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4509,6 +4568,7 @@ async fn truncation_recovery_preserves_truncated_text() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4586,6 +4646,7 @@ async fn run_finish_has_matching_thread_id() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4704,6 +4765,7 @@ async fn all_tools_suspended_pauses_run() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4765,6 +4827,7 @@ async fn completed_tool_round_clears_state_at_next_step() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4845,6 +4908,7 @@ async fn after_inference_stop_prevents_tool_execution() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4894,6 +4958,7 @@ async fn natural_end_no_tools_completes_immediately() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -4979,6 +5044,7 @@ async fn unknown_tool_in_multi_call_doesnt_crash() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5082,6 +5148,7 @@ async fn permission_denied_does_not_replay_tool() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5127,6 +5194,7 @@ async fn decision_for_unknown_call_id_returns_error() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5206,6 +5274,7 @@ async fn decision_channel_rejects_illegal_transition() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5272,6 +5341,7 @@ async fn mixed_suspended_and_completed_tools() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5475,6 +5545,7 @@ async fn parallel_tools_all_succeed() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5547,6 +5618,7 @@ async fn parallel_tools_mixed_outcomes_preserve_results() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5658,6 +5730,7 @@ async fn system_prompt_included_in_inference_request() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5737,6 +5810,7 @@ async fn message_ordering_preserved_in_inference_request() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5812,6 +5886,7 @@ async fn tool_descriptors_sent_to_llm() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5871,6 +5946,7 @@ async fn run_identity_fields_propagate_to_lifecycle() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -5972,6 +6048,7 @@ async fn context_message_suffix_system_injected() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6085,6 +6162,7 @@ async fn multiple_context_messages_injected() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6178,6 +6256,7 @@ async fn phase_hooks_fire_with_tool_call_phases() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6249,6 +6328,7 @@ async fn step_count_increments_with_tool_calls() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6298,6 +6378,7 @@ async fn token_usage_reported_in_inference_events() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6361,6 +6442,7 @@ async fn blocking_plugin_allows_non_targeted_tool() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6466,6 +6548,7 @@ async fn set_result_intercept_on_specific_tool_only() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6584,6 +6667,7 @@ async fn phase_hook_receives_tool_context() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6658,6 +6742,7 @@ async fn llm_error_on_second_step_propagates() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await;
 
@@ -6747,6 +6832,7 @@ async fn after_inference_hook_sees_llm_response() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6844,6 +6930,7 @@ async fn after_tool_execute_hook_sees_tool_result() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6896,6 +6983,7 @@ async fn max_rounds_two_stops_after_two_tool_steps() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6952,6 +7040,7 @@ async fn step_start_events_contain_step_number() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -6999,6 +7088,7 @@ async fn suspension_preserves_original_arguments() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7069,6 +7159,7 @@ async fn second_tool_not_executed_after_first_suspends() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7115,6 +7206,7 @@ async fn run_start_event_emitted_first() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7161,6 +7253,7 @@ async fn run_finish_event_emitted_last() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7220,6 +7313,7 @@ async fn tool_call_events_contain_correct_metadata() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7303,6 +7397,7 @@ async fn three_step_loop_tool_tool_response() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7358,6 +7453,7 @@ async fn lifecycle_transitions_running_to_done() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7398,6 +7494,7 @@ async fn lifecycle_transitions_running_to_waiting() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7438,6 +7535,7 @@ async fn lifecycle_transitions_running_to_done_on_cancel() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7487,6 +7585,7 @@ async fn text_delta_events_emitted_for_text_response() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7550,6 +7649,7 @@ async fn parallel_tools_have_independent_state_entries() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7605,6 +7705,7 @@ async fn parallel_tools_succeed_and_suspend_independent_states() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7658,6 +7759,7 @@ async fn parallel_tools_both_fail_independently() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7720,6 +7822,7 @@ async fn parallel_same_tool_distinct_results() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7787,6 +7890,7 @@ async fn sequential_steps_see_fresh_tool_state() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7842,6 +7946,7 @@ async fn state_snapshot_revision_increases_across_steps() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7901,6 +8006,7 @@ async fn state_snapshot_contains_extensions_with_lifecycle() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -7967,6 +8073,7 @@ async fn state_snapshot_count_matches_steps_plus_finish() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8016,6 +8123,7 @@ async fn state_snapshot_at_suspension_includes_waiting_status() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8071,6 +8179,7 @@ async fn export_persisted_after_run_has_positive_revision() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8120,6 +8229,7 @@ async fn checkpoint_store_receives_data() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8178,6 +8288,7 @@ async fn checkpoint_includes_correct_step_count() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8221,6 +8332,7 @@ async fn checkpoint_contains_state_blob() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8267,6 +8379,7 @@ async fn checkpoint_stores_thread_messages() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8352,6 +8465,7 @@ async fn checkpoint_output_supports_child_result_lookup_after_tool_messages() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8445,6 +8559,7 @@ async fn checkpoint_stores_blocked_tool_batch_consistently() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8513,6 +8628,7 @@ async fn checkpoint_stores_suspended_tool_batch_consistently() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8583,6 +8699,7 @@ async fn checkpoint_records_agent_id() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8652,6 +8769,7 @@ async fn llm_receives_all_user_messages() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8731,6 +8849,7 @@ async fn tool_results_visible_in_next_step_messages() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8836,6 +8955,7 @@ async fn context_injection_additive_not_destructive() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8901,6 +9021,7 @@ async fn token_usage_accumulates_across_steps() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -8971,6 +9092,7 @@ async fn tool_descriptors_present_even_when_unused() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9076,6 +9198,7 @@ async fn run_start_and_run_end_hooks_fire_exactly_once() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9173,6 +9296,7 @@ async fn step_start_fires_per_step() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9266,6 +9390,7 @@ async fn before_inference_hook_sees_step_count() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9371,6 +9496,7 @@ async fn plugin_context_mutation_visible_in_same_step() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9466,6 +9592,7 @@ async fn multiple_plugins_same_phase_both_fire() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9549,6 +9676,7 @@ async fn tool_result_message_contains_output() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9631,6 +9759,7 @@ async fn failed_tool_result_message_indicates_error() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9715,6 +9844,7 @@ async fn unknown_tool_result_indicates_not_found() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9771,6 +9901,7 @@ async fn tool_call_start_emitted_before_done() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9834,6 +9965,7 @@ async fn multiple_tools_each_get_start_done_pair() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9912,6 +10044,7 @@ async fn replay_tool_call_executes_original_tool() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -9968,6 +10101,7 @@ async fn replay_tool_call_executes_original_tool() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10009,6 +10143,7 @@ async fn use_decision_records_decision_payload_without_rewriting_arguments() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10074,6 +10209,7 @@ async fn pass_decision_records_decision_payload_without_rewriting_arguments() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10139,6 +10275,7 @@ async fn cancel_resume_transitions_to_cancelled_status() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10195,6 +10332,7 @@ async fn resume_with_empty_decision_result_succeeds() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10275,6 +10413,7 @@ async fn three_step_events_have_correct_overall_sequence() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10348,6 +10487,7 @@ async fn suspend_on_step_two_preserves_first_step_context() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10431,6 +10571,7 @@ async fn error_on_third_step_after_two_successful_steps() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await;
 
@@ -10494,6 +10635,7 @@ async fn mixed_tool_counts_per_step() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10540,6 +10682,7 @@ async fn full_suspend_resume_complete_lifecycle() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10594,6 +10737,7 @@ async fn full_suspend_resume_complete_lifecycle() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10647,6 +10791,7 @@ async fn inference_error_produces_error_termination() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await;
 
@@ -10732,6 +10877,7 @@ async fn token_usage_values_accumulated_across_steps() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -10932,6 +11078,7 @@ async fn awaiting_tasks_prevents_done_when_tasks_running() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -11002,6 +11149,7 @@ async fn natural_end_without_tasks_completes_normally() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -11071,6 +11219,7 @@ async fn awaiting_tasks_preserves_step_count() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -11143,6 +11292,7 @@ async fn awaiting_tasks_final_step_should_complete_once() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -11273,6 +11423,7 @@ async fn inbox_messages_injected_before_natural_end() {
         frontend_tools: Vec::new(),
         inbox: Some(inbox_rx),
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -11415,6 +11566,7 @@ async fn mid_stream_r2_recovery_injects_cancelled_tool_hint_into_next_turn() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -11542,6 +11694,7 @@ async fn mid_stream_recovery_without_parallel_tools_does_not_inject_hint() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -11668,6 +11821,7 @@ async fn malformed_tool_args_on_end_turn_injects_user_hint_for_next_turn() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
@@ -11739,6 +11893,7 @@ async fn malformed_tool_args_hint_absent_when_all_tools_have_valid_json() {
         frontend_tools: Vec::new(),
         inbox: None,
         is_continuation: false,
+        initial_state_seed: None,
     })
     .await
     .unwrap();
